[controller] Type-safe device minor allocation
- API: represent ReplicatedVolume status.deviceMinor as v1alpha1.DeviceMinor and add range validation.
- API: add DeviceMinorOutOfRangeError and exclude it from controller-gen object/deepcopy generation.
- Controller: genericize rv_controller/idpool and use IDPool[v1alpha1.DeviceMinor] for allocation.
- Codegen: regenerate deepcopy outputs.\n- Tooling: document non-API type generation hygiene in Cursor api-types rules.

+++ images/agent/internal/controllers/drbd_config/up_and_adjust_handler.go
@@ -351,7 +351,7 @@ func (h *UpAndAdjustHandler) populateResourceForNode(
 
 	vol := &v9.Volume{
 		Number:   u.Ptr(0),
-		Device:   u.Ptr(v9.DeviceMinorNumber(*h.rv.Status.DeviceMinor)),
+		Device:   u.Ptr(v9.DeviceMinorNumber(uint32(*h.rv.Status.DeviceMinor))),
 		MetaDisk: &v9.VolumeMetaDiskInternal{},
 	}
